Get the menus out of the database, it's easier to do it this way and they're not what the app's about anyhow.
At the moment the only way to switch branch is to type it into the Branch box, and the speed of time is either 0.1 or -0.1. I'm thinking of a sort of tall narrow scratchpad to fix the issue...if you're using a mouse I'll put a bar in there and pretend it's a slider.


Former-commit-id: e4afc7b3208e76ef0cf37f15c0ae73c77230323f