Some more questionable changes to touch handling
Main screen grabs every touch as long as it's visible.

You can now drag spots around even when you don't have them selected, so long as you hold them for a bit to start. This isn't really on a timer; it just takes a little while to get the list of collided widgets together, and that happens to be a nice length of time for this purpose. Feels dirty.


Former-commit-id: 05130e673007d443078e83d6b98d13d031d32744