Implement eager caching
We now fetch the state of everything in the character when it's
loaded.

It won't automagically stay up to date. Pass the name of
the character into the EngineProxy.next_tick method if you need it
up to date when the rules have run. This has some redundancy
on at least the first tick you do this, might be worth taking
out for optimization.

To deal with other kinds of time travel, you'll need to call
update_cache on everything yourself.


Former-commit-id: ec4574299f819d27d160b99371551c5cc6b57ddc