Redo the rules-handled tables
This was actually the impetus for this whole branch...
I wanted to save the state of the UI in the middle of a tick,
but couldn't do so if the UI had changed as a result of a
character_thing rule or something like that, because then
it wouldn't know what things it had dealt with already.

I guess I could have just dummied out the problematic rulebook
types? I didn't have any interface affordances for them in ELiDE
so it's not like the users would miss them.

Well, I felt I had to overhaul the database. It was a mess anyway.


Former-commit-id: 056482f5e2f017c1f00d8ba0041246ca2b63d81b