Optimize the frontend (#65)
* Reduce startup time by removing ArrowWidget from kv-lang

* Make all the bindings use triggers on the class

* Switch to fbind

* Get rid of the last bit of kv in arrow

* Remove unnecessary kv from pawnspot

* Remove unused import from arrow

* Trivially optimize the arrow point getters

Really I think I need Cython or numpy

* Use fast binding for arrow's origin and destination

Those bindings happen a lot and WeakMethod.is_dead turns
out to take a lot of time on them.

* Make SpriteBuilder a bit faster using fbind

* Fix incorrect use of unbind_uid in arrow

* Delete unused _trigger_add_arrow method

* Optimize Pallet.upd_textures

* Use a trigger to bind atlas to upd_textures

* Optimize SwatchButton and restyle a little

No more kv in this one, which turns out to be
significant, unless it was actually the reduction
in the number of distinct widgets that did it.

* Make arrow geometry less clever

The arrows now always end in the middle of one side of
a spot. It doesn't look as nice but is much more
performant when there are a lot of arrows.

* Use numpy to set arrow points rapidly on startup

Temporarily hardcoding the arrowhead size for
convenience.

* Use correct constant for 45 degree angles

* Make the partial steps of get_points into private funcs

* Let get_points return straight horiz/vert arrows again

* Remove empty on_board method from ArrowWidget

* Move even more Arrow loading into numpy

* Fix typo

* Fix typo 2

* Remove an unnecessary dictionary

* Fix some compatibility issues with recent networkx

Apparently graphs no longer have a public mapping called
`node`. I can deal...

* Remove too-clever PawnSpot.collide_point(..)

I'm not sure why this was still working on my other
machine. I haven't actually been keeping pathimgs
for some time.

* Correctly bind left-head-background arrow vertices

* Factor out one of the parts of get_points

* Bring back the old slow way of foreshortening arrows

It's reversible and reliable

* Avoid recalculating points when already done in numpy


Former-commit-id: 249e2755c46b30d7a50b5a0e793f2e093afb51ee