Permit the storage of entities that don't exist
That sounds paradoxical! It means that if you set a LiSE
entity as a stat of another entity, the stat is still valid
even if the entity it refers to has been deleted.

It's probably bad that you can still write to the database
by manipulating entities that got deleted.


Former-commit-id: 71b7d178cb4682f15f720eb2dcb6fa32f0182f13