Remove an assertion
I've decided to only consider times with actual data in them as being "loaded," meaning that games that don't have an initial keyframe will have nothing loaded, and the moment you're just starting to run the rules engine won't be loaded until it actually runs. It's normal for the present (branch, turn, tick) not to be loaded. That just means time's moving forward.


Former-commit-id: 4c8d9a33ad59e2f2215bd08b477b32582807388a