Overhaul `past` and `future` methods of `WindowDict`
Now that I'm not keeping track of the last revision anymore, it didn't make sense to permit calling these with no revision argument.

I found that, in `update_window` and `update_backward_window`, I really needed explicit control of whether or not to include the exact revision specified in the output Mapping. So I added that option, too.

I dithered on whether it was better API design to default `copy` to `True` and decided against, because I expected to set it `False` all the time, myself. Adding `__copy__` on `WindowDictPastFutureView` was an afterthought.


Former-commit-id: becaa9b7811600325e68192291449db23bcebb67