Require lxml for Lisien, so that tox can test import/export
I don't really know the best practice for "soft reqeuirements" like this. Yeah, I can mark it as an optional feature, but then you have to specifically ask for it in square brackets when you call pip, and I think most people who use Lisien will want lxml with it.


Former-commit-id: 5c8704a11d38ba92a4ec587599933bb24e242860