Stop using `OrderedDict` in `wrap`
I do want my collection types to be deterministically ordered, for iteration purposes, but I don't want them to be unequal just because the order is different. They're meant to be drop-in replacements for the standard collection types. Deterministic ordering means rules using them run the same every time, and that's all I care about.


Former-commit-id: c7793eccb95dac1df6643e68b0bdd9bdb91d5f51