Switch to fbind everywhere
Good for performance, but likely not that noticeable unless you run a load of tests at once.

I originally did this because I wanted to avoid state hanging around from previous test runs. It turned out not to totally stop it... if you use `app` in kv-lang, there's a chance it'll be a stale app from an old test. I haven't found a fix, exactly, but the workaround is in `charmenu.py` at the bottom.


Former-commit-id: 02948beff17c997379bb55b5c86fa49d6ef54075