Turn `ParquetDBLooper._iter_part_tick_to_tick` into `_list_part_tick_to_tick`
This was really only supposed to make it easier to spot a bug, but seems to make the method better behaved in general. I suspect I was advancing the iterator when I didn't mean to.

I've made it include the final tick in all cases. I don't remember making a firm decision one way or the other, and this way makes `test_export_db` pass.