Allow the user to pass their own database connector into `Engine.__init__`
And fall back to `PythonDatabaseConnector` if they don't, and we can't pick our own from the arguments.

Removes the `pack` and `unpack` arguments to the connectors' `__init__` methods, so that if a user actually does provide their own connector, we can put data in it, and interpret data from it.