Standardize the in-memory representation of `_plan_ticks`
Since the whole table is loaded and kept for the life of the app, I may as well treat it the same when it's an in-memory table in `PythonDatabaseConnector`.

This fixes an issue in `to_etree` where it was trying to load those ticks even though it already has them.