Remove most caches' names
Those names were only there to make it easier to break the debugger in the right place, back when I was trying to use the fewest distinct cache classes I could get away with. I've decided that approach was a bad idea, so now, most caches are singletons, and I can check the instance type to decide whether to break open the debugger.


Former-commit-id: e8e57ddaf17af145e4a127c37c0f434953c10b5b