Add branch, turn, and tick setters to `EngineFacade`
They won't affect the underlying `Engine`, which means that it's possible to make a bunch of changes to the facade and then time travel, at which point the facade will show the world at the destination time, but with changes meant for the original time. That could be confusing. I don't think it makes sense to just forbid time travel on `EngineFacades`, though--we need it in order to plan travel.