Normalize database connectors' initialization
The initialization happens in `__attrs_post_init__` now, and `pack` and `unpack` have to be supplied in order to even initialize the database connector.

It's no longer practical to persist the database connectors between tests; `pack` and `unpack` have lifetimes. So, when a callable is passed into `Engine` in place of a connector, we'll call it with `pack` and `unpack` to make the real connector.