Implement `OrderlyFrozenSet.__contains__`
It seems like, since `frozenset` has that method, `ABC` doesn't register that it hasn't been implemented--but since `AbstractOrderlySet` is earlier in the MRO, its inherited `__contains__` gets used instead. That `__contains__` is from the `Container` base class, which always returns `False`.

I feel like this is a problem with Python's standard `ABC`...


Former-commit-id: 165383a6817ae3ccc2203ea11b2ac29585d60548