Combine the workflows
I really meant them to run one after the other, and only when the previous step had succeeded. It didn't work, I think because when anything succeeds in a matrix, the next step sees a success and launches its own matrix.

Removed the separate step for subinterpreter tests because they just don't take that much longer...


Former-commit-id: 3b8796ffd1a8706dd9194cd6ae79b456b4916c46